fix(routing): 홈 기본 경로를 루트로 변경
