+++ docs/exec-plans/active/2026-08-05-root-home-route/delivery-state.md
@@ -1,0 +1,12 @@
+# Delivery State
+
+- status: publishing
+- branch: dev
+- base: main
+- issue:
+- pr: https://github.com/Queuing-org/frontend/pull/33
+- selected_skills: queuing-feature-delivery, queuing-ui-flow, frontend-architecture-guardrails, queuing-qa-reviewer
+- local_qa: pass — lint, 57 files/153 tests, build, diff-check, `/` 200, `/home` 308 → `/`
+- ci: pending
+- review_threads: 기존 PR 리뷰 상태 유지
+- next_action: 관련 파일만 커밋·push하고 Draft PR #33을 갱신
